refactor: rename main.js to index.js
