+++ test/test-buildSrcSet.mjs
@@ -1,6 +1,6 @@
 import md5 from 'md5';
 import assert from 'assert';
-import ImgixClient from '../src/main.mjs';
+import ImgixClient from '../src/index.mjs';
 
 function assertWidthsIncreaseByTolerance(srcset, tolerance) {
   const srcsetWidths = srcset.split(',').map((u) => {
